Fixed a bug, updated Readme.md
