+++ env_receiver.py
@@ -8,5 +8,4 @@ class EnvReceiver:
         self.token = os.getenv('BOT_TOKEN')
         self.rapid_api_host = os.getenv('RAPID_API_HOST')
         self.rapid_api_key = os.getenv('RAPID_API_KEY')
-        self.history_file = os.getenv('HISTORY_FILE')
         self.database_name = os.getenv('DATABASE_NAME')
